Move some functions to the model package
